requirements.txt
flask==2.0.1
requests==2.26.0
selenium==4.1.0
python-dotenv==0.19.0
markdown2==2.4.0
wordpress-api==1.2.9
beautifulsoup4==4.9.3
pillow==11.2.1
urllib3==1.26.20
cryptography==44.0.2
pyOpenSSL==25.0.0